+++ ChangeLog.md
@@ -10,6 +10,10 @@ Major changes:
 
 Behavior changes:
 
+* Stack uses the version of the Cabal package that comes with the specified
+  version of GHC. Stack no longer supports such Cabal versions before 2.2, which
+  came with versions of GHC before 8.4.
+
 Other enhancements:
 
 * In YAML configuration files, the `msys-environment` key is introduced to